+++ WeatherStationModel/SensorTemperaturaHumedad.h
@@ -9,11 +9,11 @@ namespace Model {
     [Serializable]
     public ref class SensorTemperaturaHumedad : public Sensor {
     public:
-        property double Temperatura;
+        property int Temperatura;
         property String^ UnidadTemp;
         property int Humedad;
         SensorTemperaturaHumedad() {}
-        SensorTemperaturaHumedad(int IdSensor, double Temperatura, String^ UnidadTemp, int Humedad);
+        SensorTemperaturaHumedad(int IdSensor, int Temperatura, String^ UnidadTemp, int Humedad);
     };
 
 }
